requirements.txt
asgiref==3.8.1
Django==4.2.11
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-jsonapi==7.0.0
inflection==0.5.1
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
